requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
python-dotenv==1.0.1
psycopg[binary]==3.2.6
Werkzeug==3.1.3
gunicorn==23.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==11.1.0
pytest==8.3.5
